requirements.txt
Babel>=1.3
decorator>=3.4.0
fixtures>=0.3.14
iso8601>=0.1.9
jsonschema>=2.0.0,<3.0.0
netaddr>=0.7.6
oslo.config>=1.2.0
paramiko>=1.9.0
pbr>=0.6,!=0.7,<1.0
pecan>=0.4.5
PrettyTable>=0.7,<0.8
PyYAML>=3.1.0
python-glanceclient>=0.9.0
python-keystoneclient>=0.8.0
python-novaclient>=2.17.0
python-neutronclient>=2.3.4,<3
python-cinderclient>=1.0.6
python-heatclient>=0.2.3
python-ceilometerclient>=1.0.6
python-ironicclient
python-muranoclient
requests>=1.1
SQLAlchemy>=0.7.8,<=0.9.99
six>=1.6.0
WSME>=0.6
